pyproject.toml
[project]
name = "liver-tumors"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "elbazhaim", email = "elbaz7haim@gmail.com" }
]

dependencies = [
    "jupyter>=1.0.0",
    "pandas>=2.2.2",
    "matplotlib>=3.9.1",
    "nibabel>=5.2.1",
    "lightning>=2.3.3",
    "torchmetrics>=1.4.0.post0",
    "seaborn>=0.13.2",
    "transformers>=4.42.4",
    "monai>=1.3.2",
    "einops>=0.8.0",
    "mlflow>=2.14.3",
    "torchvision>=0.18.1",
    "icecream>=2.1.3",
]

readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
train = 'liver_tk.train:train_model'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/liver_tk"]
